+++ CHANGELOG.md
@@ -22,6 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added support for complex logical rules via AND/OR and groups
   - Can now preview potential candidates based on the current rules selections
 - Now keeps history of reclaimed media (starting from **0.1.0-beta-14**)
+- Dashboard
+  - Reclaim history
+    - Total space reclaimed
+    - Total movies reclaimed
+    - Total series reclaimed
 
 ### Changed